backend/backend/requirements.txt
# Core Django + DRF
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.3

# JWT Authentication
djangorestframework-simplejwt>=5.2.2

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# PostgreSQL (remove if using SQLite)
psycopg2-binary>=2.9.0

# Hugging Face ecosystem
transformers>=4.41.2
datasets>=2.19.1
tokenizers>=0.19.1

# Utility
python-decouple>=3.8
dotenv

sentence-transformers
bs4
